Refactors mch domain
