+++ tests/integration/models/test_clide.py
@@ -6,7 +6,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql import text as sa_text
 
-from opencdms.models import clide
+import opencdms.provider.clide.models as clide
 from opencdms.utils.db import get_clide_connection_string
 
 DB_URL = get_clide_connection_string()
