+++ tests/integration/models/test_mch_english.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from opencdms.models.mch import english as mch_english
+from opencdms.provider.mch.models import english as mch_english
 from opencdms.utils.db import get_mch_english_connection_string
 
 DB_URL = get_mch_english_connection_string()
